refactor: expo router에 따른  app compontents들의 구조 변경 #29
- url 경로는 dir, 렌더링 파일은 index로 수정
