+++ app/(tabs)/rooms/index.tsx
@@ -2,8 +2,8 @@
 import { Fragment } from "react";
 import { View } from "react-native";
 import Top1Body10 from "../../../src/common/layout/Top1Body10";
-import RoomList from "../../rooms/_components/index/RoomList";
-import RoomCreationFloatingButton from "../../rooms/_components/index/buttons/RoomCreationFloatingButton";
+import RoomList from "../../rooms/index/_compontents/index/RoomList";
+import RoomCreationFloatingButton from "../../rooms/index/_compontents/index/buttons/RoomCreationFloatingButton";
 
 export default function RoomsScreen() {
   return (
