+++ app/prayers/[id]/_components/EmptyState.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { View, StyleSheet } from "react-native";
 import { Text } from "react-native-paper";
 import { RFValue } from "react-native-responsive-fontsize";
-import { color } from "../../../../../src/common/styles/color";
+import { color } from "../../../../src/common/styles/color";
 
 export default function EmptyState() {
   return (
