+++ app/prayers/[id]/update/index.tsx
@@ -1,15 +1,15 @@
 import React, { useState, useMemo, useEffect, useRef } from "react";
 import { StyleSheet, SafeAreaView, BackHandler } from "react-native";
-import { useSelectedPrayerTitleStore } from "../../../src/domain/prayers/stores/useSelectedPrayerTitleStore";
-import { useSelectedRoomStore } from "../../../src/domain/rooms/stores/useSelectedRoomStore";
+import { useSelectedPrayerTitleStore } from "../../../../src/domain/prayers/stores/useSelectedPrayerTitleStore";
+import { useSelectedRoomStore } from "../../../../src/domain/rooms/stores/useSelectedRoomStore";
 import { router } from "expo-router";
-import { color } from "../../../src/common/styles/color";
-import { useCloseOnBack } from "../../../src/common/services/back-handler/useCloseOnBack";
-import PrayerUpdateTop from "./_components/update/PrayerUpdateTop";
-import PrayerUpdateBody from "./_components/update/PrayerUpdateBody";
-import PrayerUpdateBottom from "./_components/update/PrayerUpdateBottom";
-import PrayerUpdateCancelDialog from "./_components/update/dialog/PrayerUpdateCancelDialog";
-import Top1Body10Bottom1 from "../../../src/common/layout/Top1Body10Bottom1";
+import { color } from "../../../../src/common/styles/color";
+import { useCloseOnBack } from "../../../../src/common/services/back-handler/useCloseOnBack";
+import PrayerUpdateTop from "./_components/PrayerUpdateTop";
+import PrayerUpdateBody from "./_components/PrayerUpdateBody";
+import PrayerUpdateBottom from "./_components/PrayerUpdateBottom";
+import PrayerUpdateCancelDialog from "./_components/dialog/PrayerUpdateCancelDialog";
+import Top1Body10Bottom1 from "../../../../src/common/layout/Top1Body10Bottom1";
 
 export default function PrayerCreationScreen() {
   const originalTitle = useSelectedPrayerTitleStore().selectedPrayerTitle;
