+++ app/rooms/[id]/_components/PrayerRoomBody.tsx
