+++ app/rooms/[id]/_components/PrayerRoomTop.tsx
